Fix: cambio de lugar los metodos de usuario a sesion
