+++ backend/src/main/java/app/servicios/ServicioSubasta.java
@@ -63,7 +63,8 @@ public class ServicioSubasta {
   public void ofertarEnSubasta(String autorId,
                                String destinoId,
                                String subastaId,
-                               List<String> rawFiguritasId) {
+                               List<String> rawFiguritasId
+  ) {
     Perfil autor = this.repositorioPerfiles.buscarPorId(autorId);
     Perfil destinatario = this.repositorioPerfiles.buscarPorId(destinoId);
     Subasta subasta = this.repoSubasta.buscarPorId(subastaId);
